+++ src/components/Map/index.tsx
@@ -1,8 +1,6 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { GoogleMap, LoadScript, Marker } from "@react-google-maps/api";
 import { locationSchema } from "@/service/package";
-import io from "socket.io-client";
-
 type Props = {
   location: locationSchema;
 };
@@ -14,23 +12,6 @@ const Map = ({ location: { lat, lng } }: Props) => {
     width: "80%",
     height: "400px",
   };
-
-  const defaultCenter = {
-    lat: 19.429995,
-    lng: -99.215664,
-  };
-
-  const [packageLocation, setPackageLocation] = useState(defaultCenter);
-  const socket = io("http://localhost:3000");
-
-  useEffect(() => {
-    socket.on("newStatus", ({ lat, lng }: any) => {
-      setPackageLocation({ lat, lng });
-    });
-    return () => {
-      socket.off("actualizacionEstado");
-    };
-  }, [socket]);
 
   return (
     <LoadScript googleMapsApiKey={apiKey}>
